+++ RELEASES.md
@@ -1,3 +1,9 @@
+v0.0.8
+
+- Renamed core.config to core.configuration
+- Added config module for config mixins
+- Added postgres config mixin with a couple of helper functions
+
 v0.0.7
 ===